create dat management library
