--- lib/persist-dat.js
@@ -1,10 +1,0 @@
-const db = require('./db')
-const encoding = require('dat-encoding')
-
-module.exports = ({ key, owner, dir }, cb) => {
-  db.put(['archive', key], {
-    key: encoding.encode(key),
-    owner,
-    dir
-  }, cb)
-}
